Make the underscore a symbol constituent
This makes it possible to use commands like M-f and M-b to move back
and forth through words in variable names suck as $foo_bar_baz, while
currently php-mode treats the variable as a single name.

The idea for this change came from an email by Deniz Dogan on the
Emacs development mailing list.