Fix font-locking for namespaces, classes, and traits
The are incorrectly given the warning face due to an incorrect
character class used in the regular expression for matching against
their names.  We will fix this part of the expression, while also
allowing backslashes to support namespaces, so that namespaces,
classes, and traits will be given the proper face.