Show anonymous functions assigned to variables on Imenu
This will catch such definitions of anonymous functions such as the
example given on PHP site itself:

    $greet = function($name)
    {
        printf("Hello %s\r\n", $name);
    };

This will show up under Imenu as ‘$greet’ under the list of Anonymous
Functions.  For now we will also relax the existing regexp for
matching functions, renaming it as ‘Named Functions’, since in that
situation we still expect the function to have an explicit name.