Add custom variable php-executable
This variable will point to the location of the PHP executable on the
user’s system.  This will allow us to do things such as integrate
Flymake support, because we need to run PHP in order for Flymake to
work correctly.  For the default value we will use a common path on
Linux where PHP is often found.