Fix bug when trying to browse local documentation
If the variable ‘php-manual-path’ is defined then
(php-search-documentation) should resort to the local documentation
instead of trying to browse the web.  However, it will do both,
opening two browser windows/tabs: one for the local documentation and
one for the documentation at the PHP website.

The (or …) expression inside of (php-search-documentation) causes this
bug.  It mistakeningly believes that (php-search-local-documentation)
returns a true value if local documentation is present, and it does
not.  So even when there is local documentation the (or …) moves on to
fetching the remote documentation.

The patch fixes this by ensuring that (php-search-local-documentation)
returns a boolean true value.