Improve argument indentation within nested functions
Currently indenting arguments in nested functions looks like so:

    $x = do_something(array(
                          'foo' => 'bar',
                          'baz' => 20
                          ));

(Yes, technically array() is not a function but a language construct.
But that distinction does not matter for our purposes.)

A more readable way to indent the arguments appeared in code posted
here on the Emacs Wiki:

    http://emacswiki.org/emacs/PhpMode#toc16

Which changes the above to look like this:

    $x = do_something(array(
             'foo' => 'bar',
             'baz' => 20
         ));

This patch copies that code, albeit with small modifications: removing
the prefix ‘ywb-’ from the functions.  This is not to slight the
original author, but to maintain some consistency in the namespace of
definitions within php-mode.  Implied in my previous statement is that
I am not the person who came up with this code.  My apologies for not
knowing who ‘ywb’ is, but he deserves proper credit for this
modification.