Make chained method calls line up in all scopes
This patch fixes a bug where method calls spread across multiple lines
were not lining up except for when they were in the global scope.
This is because we only set `c-lineup-cascaded-calls’ for calls in the
top-most section of a file.  The patch adds additional offset settings
for cc-mode to make method call alignment consistent everywhere.

Github-issue: 19