Silence compiler warnings about two variables
This patch addresses two warnings when compiling php-mode.el, related
to two variables:

1. c-vsemi-status-unknown-p
2. font-lock-syntactic-keywords

The previous commit 78cb2fc0 attempted to fix these warnings but
incorrectly used the syntax-propertize-function variable in place of
font-lock-syntactic-keywords.  As a result syntax highlighting would
no longer work and we reverted that commit with 875d33af.

This commit properly addresses the issues raised by warnings without
breaking syntax highlighting.  The fix is to use the variable
syntax-propertize-via-font-lock instead of the original attempt to use
syntax-propertize-function.  The section ‘Syntax Parsing Changes’ in
the NEWS file for Emacs 24.2 explains how this is the proper variable
to use.

Github-issue: 19
Github-issue: 24