Define the pound and newline characters as comment syntax
Currently PHP files which have comment lines like

    # This is a comment with "a string" in it

do not receive proper syntax highlighting.  The presence of any string
in such a comment, whether it uses double or single quotes, causes
Emacs to highlight the comment line with a mixture of other font-lock
faces.  This happens because Emacs does not realize that a pound sign
and a newline are comment delimeters.  We can fix this problem by
modifying the php-mode-syntax-table so that the pound sign and newline
character have the syntax properties of comment starter and comment
ender, respectively.  (See section 35.2.1 of the Elisp manual for more
details.)

Github-issue: 22