Change the array formatting example in the README
The example has each element in the array aligned by the ‘=>’
operator.  The README mentions that this alignment is not part of
php-mode; it was something I did manually.  Nonetheless, it is still
confusing to have there as it can give people the false impression
that php-mode aligns array elements in a way that it does not.  This
patch changes the example to reflect how php-mode actually indents the
array code.