Make the php-keywords list easier to read and edit
This patch breaks up the strings in the php-keywords list so that one
string appears per line, in alphabetical order.  This makes it a
little easier to see if at a glance if anything is missing, and also
decreases the chance to two patches modifying php-keywords would ever
conflict by changing the same line.

There are no logic changes in this patch.  It modifies whitespace and
layout only.