Treat ‘array’ and ‘list’ as PHP keywords
Strictly speaking neither ‘array()’ or ‘list()’ are functions.  The
PHP manual describes them as language constructs.  Since we already
treat ‘die()’ like a language construct we should be consistent and do
the same with these two.  Also, being able to write code in PHP 5.5
such as

    foreach ($users as list($firstName, $lastName)) {
        echo "First name: $firstName, last name: $lastName. ";
    }

makes it more obvious that ‘list()’ is not a function.  By treating
these two function-ish constructs like keywords we help emphasize the
special treatment they receive in PHP, which may help when users try
to do something like use ‘array()’ as a callback function and wonder
why it fails.