Update header comments for the Marmalade repo
This patch adds metadata in the header comments so that

    http://marmalade-repo.org/

will be able to read the important data from php-mode when we upload
it to that package repository.  My apologies to Turadg Aleahmad for
removing his name as the original author, but I wanted to make sure it
did not interfere with Marmalade and the parsing of the metadata.