Do not set the statement-cont offset to c-lineup-cascaded-calls
First of all, the functions which set the different coding styles
overwrite this offset value.  Second, in order to properly support
aligning method calls and the indentation support from commit 89eefdf7
we will need to write a custom line-up function for statement-cont
anyways.  So for both reasons we can delete this offset assignment.