Highlight method calls using php-function-call-face
There is already a regular expression to highlight method calls in
php-font-lock-keywords-3 but the expression for highlighting member
variables comes first in the list.  Because of their similarity the
expression for member variables always matches method calls first.  We
can fix this by simply changing the order of the expressions in the
php-font-lock-keywords-3 list so that the more specific regular
expression, the one for method calls, comes first.

Github-issue: 46