Fix annotation highlightings on the first line of C-style comments
Consider the following code:

    /* @brief This is a function.
     * @param $foo Means something.
     */

Both annotations beginning with at-signs should use the same syntax
highlighting.  However, php-mode does not highlight ‘@brief’.  This
happens because of an issue with php-annotations-inside-comment-p.
We use that macro to determine if we are inside of a comment; more
specifically, it tells use what font-face a character has at a certain
position in the buffer.  If a character is using the face
font-lock-comment-face then we assume that character is inside of a
comment, and from there we search for annotations like ‘@brief’.

The problem is that the comment in the example above a second face we
do not test for: font-lock-comment-delimiter-face.  All of the
characters before ‘@brief’ use this face, including the whitespace.
So our macro does not recognize that we are inside of a comment until
we have already reached and passed ‘@brief’, and since we only look
ahead in the buffer from that point to find annotations we never move
backwards to actually highlight ‘@brief’ properly.

We can fix this by simply making php-annotations-inside-comment-p
aware of font-lock-comment-delimiter-face.  Then our code for
annotations will recognize the start of comments earlier and correctly
highlight the example above.

Github-issue: 58