Do not highling function calls or variables within comments
This addresses the same issue as commit 4a7a6b0b.  We often use
regular expressions to capture parts of code and apply syntax
highlighting to those captured groups via font-lock definitions like
this:

    (1 font-lock-variable-name-face t t)

That third parameter causes a problem.  That controls the ‘override’
properties meaning, in this case, that Emacs will apply the face
font-lock-variable-name-face to whatever we capture even if it already
has syntax highlighting.  It overrides existing faces.  That behavior
causes comments like

    // $foo->bar

to have syntax highlighting for ‘bar’ like any other variable.  We can
fix this by changing the override parameter to ‘keep’, which tells
font-lock to keep any existing face.  Or to put it another way, to
only apply the face if none is already there.  This fixes the issue
with comments since they are highlighting before we try to
highlighting variables, function calls, et cetera.