Treat ‘encoding’ and ‘ticks’ as keywords
PHP allows the declaration of certain directives:

    http://php.net/declare

We currently highlight those directives in the warning font-face.
This patch presents a simple solution: treat the acceptable directives
as keywords.  This is not the most ideal implementation, but it is the
most simple so that we can fix the incorrect fontification we
currently have.