Use "magic" comments in PHP files to simplify indentation testing
If a line in a PHP test file contains the magic comment
`###php-mode-test### (SPEC ...)`, then each `SPEC` is executed. Each `SPEC` is
of the form `(NAME [ARG ...])` and is a test which applies to the current line
and returns non-nil if it passes. Each `ARG` is `eval`ed before being passed to
`NAME`.

A magic comment looks like this

    header('Location: http://github.com'); // ###php-mode-test### ((indent c-basic-offset))

In this case, it indicates that the current line should be indented by one
`c-basic-offset'.

Currently, the only available test is for indentation.