README: Change the absolute link to the Changelog to a relative link
Developers who create forks of php-mode may modify the Changelog to
reflect their additions, bug fixes, and so on.  But the README uses an
absolute link to the Changelog.  That means people viewing forks of
php-mode will only see a link to my copy of the Changelog, and that
will never reflect the Changelog in anyone’s fork.  Make the link
relative fixes this problem.

Admittedly this is a minuscule change.  But it is a good practice on
GitHub, where php-mode currently resides, to use relative links the
README so that forks will point to their own files.