Highlight ‘static’ as a constant when it appears in a class context
In commit 90c53729 we fixed the highlighting for ‘self’ and ‘parent’
when they appear in a class context, e.g. self::foo().  However, we
forgot to treat ‘static’ in the same way.  When we have code like

    static::foo();

we want ‘static’ to appear in the constant font face because that is
consistent with how we treat the other two words.  This patch gives
‘static’ the proper face in that situation.  Note that we must tell
font lock mode to override any existing face for ‘static’ or otherwise
it will not work since we apply faces to it for other reasons.