Associate file extensions when installing from an Elisp repository
If a user installs php-mode from a repository, e.g. MELPA, then Emacs
will not recognize the file extensions for PHP.  This happens because
we add those extensions to ‘auto-mode-alist’ via code in a ‘defcustom’
definition.  However, there is no guarantee that Emacs will execute
that code when we install php-mode from a repository.

This patch removes that ‘defcustom’ code and replaces it with a loop
that will always add the appropriate extensions to ‘auto-mode-alist’.
Because of this change users can no longer customize the valid
extensions for PHP files, but it is worthwhile to temporarily remove
that customization in order to fix this more important problem.

Github-issue: 79