Correct indentation for array arguments (tests/issues-14.php)
Commit 055d0810 began a transition to using more of cc-mode’s
machinery for customizing indentation.  In the process this broke the
indentation we expect from the ‘tests/issues-14.php’ script.  This
failure went unnoticed because we expected the test to fail, and so
every time we would run the test suite the test would successfully
fail.  After the code was in place to make indentation work properly
we failed to update the test accordingly so that it would no longer
expect a failure.

Due to this mismanagement of the test suite on my part, a regression
which re-broke indentation entered the master branch unnoticed.

This patch corrects the problem in a number of ways:

1. We no longer expect ‘php-mode-test-issue-14’ to fail.  When we run
the test suite it must pass.

2. We explicitly use the ‘php-lineup-arglist-intro’ and
‘php-lineup-arglist-close’ functions for the cc-mode offsets
‘arglist-intro’ and ‘arglist-close’, respectively.  We use these in
all three styles.

3. When loading php-mode we explicitly enable the user’s coding style
before running associated hooks.

With this patch in place we can start a fresh instance of ‘emacs -Q’,
load php-mode, and get the expected indentation in the issue-14 test
for all three coding styles we define: PEAR, Drupal, and WordPress.

Finally, this patch performs some reformatting of the Elisp itself for
readability and consistency.

Github-issue: 80