Do not trigger search error in php-mode-test-issue-19
We search for the string ‘$object->’ to determine if its position
matches the arrow on the next line.  Because we test this in different
contexts (e.g. in methods, in closures, at the top level) we search
for the string in a loop.  When we reach the end of the buffer,
however, search-forward returns an error because there are no more
‘$object->’ strings to find.  This patch changes the search so it does
not raise the error but only returns nil at the end, causing the loop
to end normally.

Github-issue: 81