Fix chained method alignment
Commit f933ec2d removed the cc-mode setting for ‘statement-cont’ but
we never re-introduced it when implementing the different PHP styles.
That breaks the alignment of chained methods across different lines.
This patch fixes this by re-introducing ‘statement-cont’ to each style
and by giving ‘brace-list-intro’ and ‘brace-list-entry’ proper values
so that methods are correctly aligned in the different contexts
presented in the issue-19 test.

This patch causes php-mode-test-issue-19 to pass successfully.

Github-issue: 81