README: Document how chained method alignment may fail
Because PHP mode uses CC mode for indentation a user may be using a
variety of styles that we cannot anticipate.  We implement chained
method alignment, e.g

    $object->foo()
           ->bar()
           ->baz();

by defining certain properties for the three indentation styles that
are special to PHP mode.  We cannot make these same modifications to
other styles because that would overstep the bounds of PHP mode; it
should not affect the behavior of other modes that also use CC mode
for indenting code in unrelated languages.

Therefore we need to make it clear to the user that method alignment
is only guaranteed to work if he or she is using one of the
indentation styles specific to PHP mode.

Github-issue: 81