List all methods via Imenu regardless of their visibility
To find a method using Imenu the programmer must remember its
visibility, e.g. if it was a ‘public’ or ‘private’ method, because we
group methods by their visibility.  If the user does not remember the
visibility then he must look through all of the sub-menus.

This patch adds an ‘All Methods’ entry to the Imenu list that contains
every method in a file, regardless of its visibility.  This makes it
possible for programmers to find methods more easily if they only
remember the name and not the visibility.