README: Correct the documentation for chained-method call alignment
The current documentation says that PHP Mode only tries to guarantee
proper alignment when using one of three coding styles:

1. WordPress
2. Drupal
3. Pear

However, Michael Dwyer’s recent addition of a Symfony2 coding style is
another which should perform proper alignment.  So instead of
explicitly listing each coding style it is better to make note in the
README that chained-method call alignment is only a guarantee when
using the styles described under the ‘Coding Styles’ section.  This
way additions of new coding styles in the future will not require any
further changes to this part of the README.