Document Subword Mode in the README
The README briefly explains the purpose of Subword Mode, linking to
the official manual for more information.  PHP Mode does not enable
Subword Mode by default.  So the README describes two ways to use it:

1. We can add a hook to PHP Mode that enables Subword Mode.  The
README demonstrates an example of this.

2. We can toggle the minor mode on and off by using ‘C-c C-w’ in PHP
Mode (and other modes based on cc-mode).

Adding (global-subword-mode t) to an Emacs config file will enable
Subword Mode for all buffers.  The README omits this third method
because, based on personal experience, I do not want the mode
affecting every buffer and file I use.  Instead I find it easier to
setup a hook or toggling the mode, as described in the README.

GitHub-Issue: 89