Fix indentation of statements after ‘foreach’ without braces
PHP Mode fails to indent the following code:

    foreach ($x as $s)
        echo $s;

It does not indent the second line if there are no braces following
the ‘foreach’.  Commit 89eefdf7 previously fixed this problem but a
regression happened when making changes to all of the indentation
styles.  Each CC Mode style that PHP Mode provides defines a value for
‘statement-cont’.  That controls the second line in the example.

We changed ‘statement-cont’ in the three original styles to use the
lineup function ‘c-lineup-cascaded-calls’ because that tells PHP Mode
to properly indent cascaded method calls, e.g.:

    foreach ($x as $s)
        echo $s->foo()
               ->bar();

However, changing the value of ‘statement-cont’ broke indentation for
the first example.  We address this problem by redefining
‘statement-cont’ in every style except ‘Symfony2’.  Each style
attempts to first use ‘c-lineup-cascaded-calls’ when indenting a line
that falls under the control of ‘statement-cont’, but if that returns
nil, which happens in the first ‘foreach’ above, then the style
defaults to indenting by ‘c-basic-offset’.  In other words, we first
attempt to align the code as a chained method, and failing that we
simply indent it normally.

This patch omits any change to the ‘Symfony2’ style because it uses a
different function for ‘statement-cont’ than all of the rest and will
require separate testing and modification.

Finally, this patch also introduces a unit-test for this problem so
that we can avoid future regressions.  The test contains the first
example in the commit message and makes sure that PHP indents the
second line by the value described by ‘c-basic-offset’.

GitHub-Issue: 99