Use mode-specific functions for defun movement
We derive PHP Mode from CC Mode.  So by default PHP Mode inherits
these bindings:

1. C-M-a    c-beginning-of-defun

2. C-M-e    c-end-of-defun

This patch replaces those functions with php-beginning-of-defun and
php-end-of-defun, respectively, so that Emacs uses them when people
are using PHP Mode.  We remap the functions themselves and not their
key-bindings so that Emacs will use those two PHP-specific functions
for any keys which the user binds to the default CC Mode commands.

Special thanks to Tom Willemse for providing this idea.  This patch
also makes a correction in the README where I had mistakenly
misspelled his name.

GitHub-Issue: 116