Fix regex for namespaces in function calls
The regular expression we use to apply syntax highlighting to
namespaces in function calls breaks when it comes to sub-namespaces.
That means that PHP Mode highlights

    \Foo\bar();

correctly but it messes up

    \Foo\Bar\baz();

The proper font-face is only applied to the inner-most namespace, i.e
‘\Bar’ instead of the full ‘\Foo\Bar’.  This patch corrects that
problem so that the full namespace has the correct face.

GitHub-Issue: 26
Signed-off-by: Eric James Michael Ritz <lobbyjones@gmail.com>