Stop marching indentation for try-catch blocks
This patch not only fixes the issue but also includes a test case
which has a lot of commentary explaining how CC Mode tries to indent
try-case blocks and the potential problems therein.  Our specific fix
in this patch is to add the keyword `catch' to the constant list
`php-block-stmt-2-kwds'.  The result is that it adds the keyword to
the CC Mode list `c-block-stmt-2-key' and that automatically takes
care of indenting try-catch blocks in situations where we put the
keywords and braces on the same lines together.

GitHub-Issue: 68