GitHub-Issue: 115 (Unit Test)
We expect this test to fail because of a current limitation in the
test suite.  In order to test for indentation we use ‘magic comments’
and describe the correct, expected offset for indentation that we want
to see when running the tests.  The test for issue 115 has this:

    // ###php-mode-test### ((indent c-lineup-cascaded-calls))

The test suite code will fail on this magic comment because it tries
to treat `c-lineup-cascaded-calls' like a variable that we can compare
to the indentation.  The problem is that `c-lineup-cascaded-calls' is
a function, and we need to apply it to the correct ‘language element’
in order to get an indentation we can compare against.

Until the test suite supports this behavior we will expect the test
for issue 115 to always fail.