Use the PEAR indentation style by default for unit tests
We can tell the `with-php-mode-test' macro to use an explicit
indentation style if necessary.  However, we rarely do so, meaning
that the temporary buffers we create for testing indentation take on
the user’s preferred default indentation style, e.g. `gnu' or `linux'.
We only ensure correct indentation for the styles that PHP Mode
explicitly provides to users.  So this patch changes the behavior of
the macro to use the PEAR style of indentation by default, that way we
can be sure we are testing indentation with a style we support.

This patch also modifies the macro so that it enables PHP Mode before
selecting any style, as the original order of trying to select the
style first can cause errors.

This change relates to the issue below, as its unit test would no
longer pass since I was not using one of PHP Mode’s specific styles.

GitHub-Issue: 29