Align chained method calls inside of arrays
This patch modifies the four PHP Mode specific indentation styles so
that method calls in arrays align like so:

    $x = ["x" => $this->foo()
                      ->bar()
                      ->baz()];

Admittedly this may introduce unexpected indentation in some
circumstances, particularly with the Drupal style, as it was the only
style that defined `arglist-cont-nonempty'.  We do not have any tests
which use that cc-mode syntactic symbol in other contexts.  So this
patch implements the indentation above but because our test suite does
not thoroughly examine `arglist-cont-nonempty' this patch should be
first on our list of ‘things to check’ and possibly revert if we
receive any bug reports related to `arglist-cont-nonempty'.

GitHub-Issue: 115