Line-up multiple namespaces in a multi-line ‘use’ statement
This patch modifies all four PHP Mode styles of indentation so that
code such as

    use ReflectionMethod,
        ReflectionObject,
        ReflectionProperty;

will align the namespaces.  We do this by changing how each style
handles the ‘topmost-into-cont’ language element.  That is the
syntactical element that CC Mode uses for the second and third lines
in the example above.  Normally we want to use the function
‘c-lineup-cascaded-calls’ for those elements.  With this patch we
still try to use that function first; however, if it performs no
indentation then we attempt to use ‘php-lineup-arglist-intro’, as that
will align the namespaces.

GitHub-Issue: 100