Fix ‘final’ not highlighted as a keyword in some methods
Consider the following code:

    class MyClass
    {
        public final function foo()
        {
            return null;
        }
    }

PHP Mode does not use the keyword font-face for ‘final’ in that code.
This patch corrects that bug by adding ‘final’ to the list of PHP
keywords (along with alphabetizing them).  It also removes some
unnecessary re-highlighting that we do, highlighting which is already
covered by the aforementioned list of keywords.

GitHub-Issue: 129
Signed-off-by: Eric James Michael Ritz <lobbyjones@gmail.com>