Improve array indentation for the PEAR style
This patch modifies the style so that it will attempt to use multiple
line-up functions (from CC Mode) in order to align the keys of arrays.
Using the ‘c-lineup-arglist-intro-after-paren’ function lets us
correctly align those arrays.  However, it breaks indentation for some
test scripts, e.g. the test for issue 29.  So this is a step in the
right direction but we must fix the buggy side-effects.

GitHub-Issue: 130