Append file patterns to auto-mode-alist
Some other modes stake a claim to PHP files, e.g. mmm-mode.  PHP Mode
blasts over auto-mode-alist and depending on the order in which a user
loads each PHP Mode may usurp association with PHP files from the
other modes, and that can be annoying to users of those other modes.
This patch appends the file patterns for PHP Mode to the end of
auto-mode-alist so that they will not over-shadow any existing
associations which may already exist.

GitHub-Issue: #32
GitHub-Issue: #105
GitHub-Issue: purcell/mmm-mode#32