Prepend ‘file://’ when searching local documentation
Some browsers, e.g. Emacs’ own new eww, require the ‘file://’ prefix
before the value of php-manual-path whenever we are searching a local
copy of the documentation.  Some browsers do not need the protocol
prefix but it is never incorrect to add it if not already there.

This patch also adds phil-s to the list of contributors since we use
code from his bug report on this issue.

GitHub-Issue: Fixes #146