Restore broken psr2 and symfony2 styles
Those two were removed in the first cc-mode conversion commit. Below a
short summary of the styles and how I intend them to behave in this
branch.

Drupal: Sets indentation to two spaces, some indentation tweaks. Does
automatic removal of trailing whitespace.

Wordpress: Indentation tweaks and tab-width of four spaces.

PSR2, Symfony2 and PEAR: They are basically the same. The Symfony2 style
allows the unofficial 'hanging semicolon' behaviour. PEAR is the
default. I think it would be better for php-mode to just forget about
PEAR, drop Symfony2 as a style (the framework uses PSR-2) and use PSR-2
as default. The hanging semicolon could be a configurable option,
independent of the style in effect.

This commit fixes the failing test on whitespace removal (issue 53).