Better highlighting of variable and array-keyword
Added test cases for the following:

* all words in $object->variable should get the variable face
* array should get a keyword face
* except when used in casts like this (array)

This behaviour is already present in mainline php-mode, added it back
here because it seems sane/nice and should not be removed. Having
regexpes to do some highlighting is fine for cases that cc-mode does not
provide.

Also fixed: unset should not get a type face