Better configuration of multiline strings
cc-mode was configured to have escaped newlines in string literals and
to not have multiline strings, while it's actually the other way
around. This caused the opening " of a multiline string to get a
warning-face. I changed the test for issue-21 so it will fail in that
case. The test still passes because this commit fixes the issue.

The behaviour of escaped quotes as described in issue-174 is now
different, but not yet fixed.