Fix bad highlighting after escaped quotation mark
For keyword fontification to work we need backslashes as part of
words. But when doing syntactic fontification (comments and strings)
backslashes must have escape syntax.

I considered two ways of dealing with this:

1. Make \ word syntax by default and advice
`font-lock-fontify-syntactically-region' to temporarily set it to escape.

2. Make \ escape syntax by default and advice
`font-lock-fontify-keywords-region' to temporarily set it to word.

This commit implemented the second option so we don't change the
behaviour of forward-word compared to current stable php-mode.