Add missing language constructs to php-mode
A new test contains all language constructs and reserved keywords from
the PHP 5.6 manual. The test checks that they all get the
keyword-face. In addition to that, for keywords that preceed a type
identifier like instanceof and insteadof, it is checked that the
identifier gets the type-face.

Some keywords in c-other-kwds were already specified in another *-kwds,
those are now removed. Keywords added in this commit are: insteadof,
callable (the type hint now works just like 'array') and __halt_compiler().

This resolves issue #180.