README: Clarify a few things
1. Shorten the ’Status’ section to further emphasize that PHP Mode is
not dead even though I have stepped back to more of a co-maintainer
role.  This change is motivated by a few comments on a forum where
users mistakeningly thought the project was dead and I was no longer
accepting patches.

2. Mention that people can link to personal websites as contributors
instead of linking to their GitHub page, if they prefer.

3. Explain that patches should only update `php-mode-modified` if they
also affect the `php-mode.el` file itself.  If they do not change that
file then it is not necessary to update the constant.