[Fix] Correct indentation of closures containing the `use` keyword
Consider the following code, with line numbers:

    1: array_map(function ($x, $y) {
    2:     return $x + $y;
    3: }, $foo);

PHP Mode indents this code "correctly", i.e. the way most users would
expect.  Specifically, PHP Mode indents line 2 per the settings for
`defun-block-intro` and line 3 per `inline-close`.  This is the
desired behavior.

Now consider when we introduce `use` into the closure:

    1: array_map(function ($x, $y) use ($z) {
    2:             return $x + $y;
    3:     }, $foo);

We end up with something like this.  What we want is for PHP Mode to
indent this code in the same way as the first example.  This patch
addresses that problem.

First we add `use` to `c-lambda-kwds`.  This causes CC Mode to add the
`inlambda` syntax context to lines 2 and 3, which `C-c C-s` will
demonstrate.  In our definition for `c-offset-alist` we have...

    (inlambda . 0)

...but without the modification to `c-lambda-kwds` that indentation is
never applied.

Second, we introduce a new offset:

    (lambda-intro-cont . +)

This offset handles the following type of code:

    1: array_map(
    2:    function ($x, $y)
    3:        use ($z) {
    4:            return $x + $y;
    5:        },
    6:    $foo);

Here CC Mode indents line 2 according to `lambda-intro-cont`.
Admittedly this is not a perfect solution, because it creates
undesirable indentation if we combined lines 1 and 2 in the example
code above.  This patch tries to do "The Closest to Right Thing"
possible in that scenario, but the behavior remains a undesirable
compromise at best.

The `use` keyword appears in two other situations in PHP:

1. Namespaces, e.g.:

    use Foo\Bar as Baz;

2. Class traits, e.g.:

    class Foo extends Bar {
        use Baz;
    }

This patch does not affect the indentation in either situation because
CC Mode does not apply `inlambda` to the lines containing `use`.

Finally the patch adds a new test for the relevant GitHub issue.

Fixes: #259
Reviewed-By: Syohei Yoshida
Special-Thanks: Andrea Turso
Special-Thanks: James W. Dunne