[Feature] Highlight return types in functions and methods
Consider the following:

    function add(int $x, int $x) : int
    {
        return $x + $y;
    }

PHP Mode highlights the first instances of `int` using
`font-lock-type-face`, but does nothing with the third instance of
`int`, i.e. the return type of the function.  This patch modifies the
list of font-lock keywords so that every use of `int` in the example
above receives the same type of syntax highlighting.  This also works
with class methods like so:

    class User
    {
        private $address;

        public function getAddress() : Address
        {
            return $this->address;
        }
    }

The implementation is "dumb" in the sense that no semantic analyzation
takes place.  All we do is use a rather simple regular expression to
look for patterns like...

    function [followed by anything] : [type]

...to perform the highlighting.  In the future, as PHP Mode
incorporates more intelligent, semantic-aware processing of code, we
will most likely want to come back and remove this highlighting hack.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>