[Docs] Update various information in the README
This patch modifies the README by...

1. Explicitly listing the current maintainers, because some people see
old comments and blog posts of mine (Eric) in which I talk about
finding a co-maintainer for PHP Mode.  However, Syohei Yoshida is the
co-maintainer and there is currently no need for another.  So this
patch lists us as maintainers in the README so that people are clear
on the status of PHP Mode as it pertains to maintenance.

2. Stresses using GitHub issues for bug reports and feature requests
as opposed to emailing them directly to me.  Again, this is because
PHP Mode now has two maintainers.  Creating issues on GitHub
significantly increases the chances of a timely response.

3. Briefly describes the on-going work to incorporate CEDET, EDE,
Semantic, and other functionality to make PHP Mode behave more like a
"traditional IDE" for users who want that.

4. Simplifies and clarifies the description of our process for
reviewing and merging pull-requests on GitHub.