[Fix] Remove global advice around font-lock function for keywords
Commit 55676e80 ("Fix bad highlighting after escaped quotation mark")
introduced `defadvice` around `font-lock-fontify-keywords-region` to
fix a problem with syntax highlighting, particularly as it relates to
namespaces in PHP.  However, this approach causes problems with other
modes that use that function, namely...

1. `rainbow-delimiters-mode`
2. `lisp-extra-font-lock-mode`

...particularly the latter.

This patch is the first step in addressing the problem: removing the
`defadvice` entirely.  As pointed out by Fanael Linithien, it is a bad
idea for PHP Mode to globally define advice for a Font Lock function
likely to be used by a number of other modes.

This patch introduces a regression by re-introducing the problem that
55676e80 fixed, but by removing the `defadvice` we are paving the way
to fix that problem in a different way that will not conflict with
other modes by not defining advice around any Font Lock functions.

GitHub-Issue: #280
Special-Thanks: Fanael Linithien
Special-Thanks: Sebastian Fieber